renamed activemq to activemq-classic
