+++ docs-examples/build.gradle
@@ -14,7 +14,7 @@ subprojects { Project subproject ->
     dependencies {
         compileOnly 'org.graalvm.nativeimage:svm'
         testImplementation project(':jms-test')
-        testImplementation project(':jms-activemq')
+        testImplementation project(':jms-activemq-classic')
         testImplementation 'io.micronaut.reactor:micronaut-reactor'
     }
