+++ docs-examples/example-groovy/src/test/groovy/io/micronaut/jms/docs/quickstart/TextConsumer.groovy
@@ -5,7 +5,7 @@ import io.micronaut.jms.annotations.JMSListener
 import io.micronaut.jms.annotations.Queue
 import io.micronaut.messaging.annotation.Body
 
-import static io.micronaut.jms.activemq.configuration.ActiveMqConfiguration.CONNECTION_FACTORY_BEAN_NAME
+import static io.micronaut.jms.activemq.classic.configuration.ActiveMqClassicConfiguration.CONNECTION_FACTORY_BEAN_NAME
 // end::imports[]
 import io.micronaut.context.annotation.Requires
