+++ settings.gradle
@@ -1,7 +1,7 @@
 rootProject.name = 'jms'
 
 include 'jms-core'
-include 'jms-activemq'
+include 'jms-activemq-classic'
 include 'jms-activemq-artemis'
 include 'jms-sqs'
 include 'jms-test'
